requirements.txt
Package            Version
------------------ ---------
beautifulsoup4     4.12.3
bs4                0.0.2
certifi            2024.8.30
charset-normalizer 3.4.0
greenlet           3.1.1
idna               3.10
mutagen            1.47.0
pip                22.0.2
playwright         1.48.0
pyee               12.0.0
PyQt6              6.7.1
PyQt6-Qt6          6.7.3
PyQt6_sip          13.8.0
python-dotenv      1.0.1
requests           2.32.3
setuptools         59.6.0
soupsieve          2.6
typing_extensions  4.12.2
urllib3            2.2.3
